+++ test/warnings.js
@@ -1,8 +1,7 @@
-"use strict";
-var sinon = require("sinon");
-var setCookie = require("../lib/set-cookie.js");
+import sinon from "sinon";
+import setCookie from "../lib/set-cookie.js";
 
-describe("set-cookie-parser", function () {
+describe("warnings", function () {
   var sandbox = sinon.createSandbox();
 
   afterEach(function () {